Isolate class and module functions into their own runtime modules
