+++ opal/runtime/boot.js
@@ -185,6 +185,7 @@
   Opal.slice = $slice;
   Opal.splice = $splice;
   Opal.has_own = $has_own;
+  Opal.set_proto = $set_proto;
 
   // Helpers
   // -----
@@ -256,23 +257,6 @@
     }
   }
 
-  // TracePoint support
-  // ------------------
-  //
-  // Support for `TracePoint.trace(:class) do ... end`
-  Opal.trace_class = false;
-  Opal.tracers_for_class = [];
-
-  function invoke_tracers_for_class(klass_or_module) {
-    var i, ii, tracer;
-
-    for(i = 0, ii = Opal.tracers_for_class.length; i < ii; i++) {
-      tracer = Opal.tracers_for_class[i];
-      tracer.trace_object = klass_or_module;
-      tracer.block.$call(tracer);
-    }
-  }
-
   // Constants
   // ---------
   //
@@ -285,39 +269,8 @@
   // - nesting: the namespaces wrapping the current scope, e.g. nesting inside
   //            `module A; module B::C; end; end` is `[B::C, A]`
 
-  // Get the constant in the scope of the current cref
-  function const_get_name(cref, name) {
-    if (cref) {
-      if (cref.$$const[name] != null) { return cref.$$const[name]; }
-      if (cref.$$autoload && cref.$$autoload[name]) {
-        return Opal.handle_autoload(cref, name);
-      }
-    }
-  }
-
-  Opal.const_get_name = const_get_name;
-
   // Initialize the top level constant cache generation counter
   Opal.const_cache_version = 1;
-
-  // Walk up the ancestors chain looking for the constant
-  function const_lookup_ancestors(cref, name) {
-    var i, ii, ancestors;
-
-    if (cref == null) return;
-
-    ancestors = $ancestors(cref);
-
-    for (i = 0, ii = ancestors.length; i < ii; i++) {
-      if (ancestors[i].$$const && $has_own(ancestors[i].$$const, name)) {
-        return ancestors[i].$$const[name];
-      } else if (ancestors[i].$$autoload && ancestors[i].$$autoload[name]) {
-        return Opal.handle_autoload(ancestors[i], name);
-      }
-    }
-  }
-
-  Opal.const_lookup_ancestors = const_lookup_ancestors;
 
   // Register the constant on a cref and opportunistically set the name of
   // unnamed classes/modules.
@@ -484,91 +437,6 @@
   };
   Opal.allocate_class = $allocate_class;
 
-
-  function find_existing_class(scope, name) {
-    // Try to find the class in the current scope
-    var klass = const_get_name(scope, name);
-
-    // If the class exists in the scope, then we must use that
-    if (klass) {
-      // Make sure the existing constant is a class, or raise error
-      if (!klass.$$is_class) {
-        $raise(Opal.TypeError, name + " is not a class");
-      }
-
-      return klass;
-    }
-  }
-
-  function ensureSuperclassMatch(klass, superclass) {
-    if (klass.$$super !== superclass) {
-      $raise(Opal.TypeError, "superclass mismatch for class " + klass.$$name);
-    }
-  }
-
-  Opal.klass = function(scope, superclass, name) {
-    var bridged;
-
-    if (scope == null || scope === '::') {
-      // Global scope
-      scope = _Object;
-    } else if (!scope.$$is_class && !scope.$$is_module) {
-      // Scope is an object, use its class
-      scope = scope.$$class;
-    }
-
-    // If the superclass is not an Opal-generated class then we're bridging a native JS class
-    if (
-      superclass != null && (!superclass.hasOwnProperty || (
-        superclass.hasOwnProperty && !superclass.hasOwnProperty('$$is_class')
-      ))
-    ) {
-      if (superclass.constructor && superclass.constructor.name == "Function") {
-        bridged = superclass;
-        superclass = _Object;
-      } else {
-        $raise(Opal.TypeError, "superclass must be a Class (" + (
-          (superclass.constructor && (superclass.constructor.name || superclass.constructor.$$name)) ||
-          typeof(superclass)
-        ) + " given)");
-      }
-    }
-
-    var klass = find_existing_class(scope, name);
-
-    if (klass != null) {
-      if (superclass) {
-        // Make sure existing class has same superclass
-        ensureSuperclassMatch(klass, superclass);
-      }
-    }
-    else {
-      // Class doesn't exist, create a new one with given superclass...
-
-      // Not specifying a superclass means we can assume it to be Object
-      if (superclass == null) {
-        superclass = _Object;
-      }
-
-      // Create the class object (instance of Class)
-      klass = $allocate_class(name, superclass);
-      $const_set(scope, name, klass);
-
-      // Call .inherited() hook with new class on the superclass
-      if (superclass.$inherited) {
-        superclass.$inherited(klass);
-      }
-
-      if (bridged) {
-        Opal.bridge(bridged, klass);
-      }
-    }
-
-    if (Opal.trace_class) { invoke_tracers_for_class(klass); }
-
-    return klass;
-  };
-
   // Define new module (or return existing module). The given `scope` is basically
   // the current `self` value the `module` statement was defined in. If this is
   // a ruby module or class, then it is used, otherwise if the scope is a ruby
@@ -613,223 +481,6 @@
     return module;
   };
   Opal.allocate_module = $allocate_module;
-
-  function find_existing_module(scope, name) {
-    var module = const_get_name(scope, name);
-    if (module == null && scope === _Object) module = const_lookup_ancestors(_Object, name);
-
-    if (module) {
-      if (!module.$$is_module && module !== _Object) {
-        $raise(Opal.TypeError, name + " is not a module");
-      }
-    }
-
-    return module;
-  }
-
-  Opal.module = function(scope, name) {
-    var module;
-
-    if (scope == null || scope === '::') {
-      // Global scope
-      scope = _Object;
-    } else if (!scope.$$is_class && !scope.$$is_module) {
-      // Scope is an object, use its class
-      scope = scope.$$class;
-    }
-
-    module = find_existing_module(scope, name);
-
-    if (module == null) {
-      // Module doesnt exist, create a new one...
-      module = $allocate_module(name);
-      $const_set(scope, name, module);
-    }
-
-    if (Opal.trace_class) { invoke_tracers_for_class(module); }
-
-    return module;
-  };
-
-  // Return the singleton class for the passed object.
-  //
-  // If the given object alredy has a singleton class, then it will be stored on
-  // the object as the `$$meta` property. If this exists, then it is simply
-  // returned back.
-  //
-  // Otherwise, a new singleton object for the class or object is created, set on
-  // the object at `$$meta` for future use, and then returned.
-  //
-  // @param object [Object] the ruby object
-  // @return [Class] the singleton class for object
-  Opal.get_singleton_class = function(object) {
-    if (object.$$is_number) {
-      $raise(Opal.TypeError, "can't define singleton");
-    }
-    if (object.$$meta) {
-      return object.$$meta;
-    }
-
-    if (object.hasOwnProperty('$$is_class')) {
-      return Opal.build_class_singleton_class(object);
-    } else if (object.hasOwnProperty('$$is_module')) {
-      return Opal.build_module_singleton_class(object);
-    } else {
-      return Opal.build_object_singleton_class(object);
-    }
-  };
-
-  // helper to set $$meta on klass, module or instance
-  function set_meta(obj, meta) {
-    if (obj.hasOwnProperty('$$meta')) {
-      obj.$$meta = meta;
-    } else {
-      $prop(obj, '$$meta', meta);
-    }
-    if (obj.$$frozen) {
-      // If a object is frozen (sealed), freeze $$meta too.
-      // No need to inject $$meta.$$prototype in the prototype chain,
-      // as $$meta cannot be modified anyway.
-      obj.$$meta.$freeze();
-    } else {
-      $set_proto(obj, meta.$$prototype);
-    }
-  };
-
-  // Build the singleton class for an existing class. Class object are built
-  // with their singleton class already in the prototype chain and inheriting
-  // from their superclass object (up to `Class` itself).
-  //
-  // NOTE: Actually in MRI a class' singleton class inherits from its
-  // superclass' singleton class which in turn inherits from Class.
-  //
-  // @param klass [Class]
-  // @return [Class]
-  Opal.build_class_singleton_class = function(klass) {
-    if (klass.$$meta) {
-      return klass.$$meta;
-    }
-
-    // The singleton_class superclass is the singleton_class of its superclass;
-    // but BasicObject has no superclass (its `$$super` is null), thus we
-    // fallback on `Class`.
-    var superclass = klass === BasicObject ? Class : Opal.get_singleton_class(klass.$$super);
-
-    var meta = $allocate_class(null, superclass, true);
-
-    $prop(meta, '$$is_singleton', true);
-    $prop(meta, '$$singleton_of', klass);
-    set_meta(klass, meta);
-    // Restoring ClassName.class
-    $prop(klass, '$$class', Opal.Class);
-
-    return meta;
-  };
-
-  Opal.build_module_singleton_class = function(mod) {
-    if (mod.$$meta) {
-      return mod.$$meta;
-    }
-
-    var meta = $allocate_class(null, Opal.Module, true);
-
-    $prop(meta, '$$is_singleton', true);
-    $prop(meta, '$$singleton_of', mod);
-    set_meta(mod, meta);
-    // Restoring ModuleName.class
-    $prop(mod, '$$class', Opal.Module);
-
-    return meta;
-  };
-
-  // Build the singleton class for a Ruby (non class) Object.
-  //
-  // @param object [Object]
-  // @return [Class]
-  Opal.build_object_singleton_class = function(object) {
-    var superclass = object.$$class,
-        klass = $allocate_class(nil, superclass, true);
-
-    $prop(klass, '$$is_singleton', true);
-    $prop(klass, '$$singleton_of', object);
-
-    delete klass.$$prototype.$$class;
-
-    set_meta(object, klass);
-
-    return klass;
-  };
-
-  // Returns an object containing all pairs of names/values
-  // for all class variables defined in provided +module+
-  // and its ancestors.
-  //
-  // @param module [Module]
-  // @return [Object]
-  Opal.class_variables = function(module) {
-    var ancestors = $ancestors(module),
-        i, length = ancestors.length,
-        result = {};
-
-    for (i = length - 1; i >= 0; i--) {
-      var ancestor = ancestors[i];
-
-      for (var cvar in ancestor.$$cvars) {
-        result[cvar] = ancestor.$$cvars[cvar];
-      }
-    }
-
-    return result;
-  };
-
-  // Sets class variable with specified +name+ to +value+
-  // in provided +module+
-  //
-  // @param module [Module]
-  // @param name [String]
-  // @param value [Object]
-  Opal.class_variable_set = function(module, name, value) {
-    var ancestors = $ancestors(module),
-        i, length = ancestors.length;
-
-    for (i = length - 2; i >= 0; i--) {
-      var ancestor = ancestors[i];
-
-      if ($has_own(ancestor.$$cvars, name)) {
-        ancestor.$$cvars[name] = value;
-        return value;
-      }
-    }
-
-    module.$$cvars[name] = value;
-
-    return value;
-  };
-
-  // Gets class variable with specified +name+ from provided +module+
-  //
-  // @param module [Module]
-  // @param name [String]
-  Opal.class_variable_get = function(module, name, tolerant) {
-    if ($has_own(module.$$cvars, name))
-      return module.$$cvars[name];
-
-    var ancestors = $ancestors(module),
-      i, length = ancestors.length;
-
-    for (i = 0; i < length; i++) {
-      var ancestor = ancestors[i];
-
-      if ($has_own(ancestor.$$cvars, name)) {
-        return ancestor.$$cvars[name];
-      }
-    }
-
-    if (!tolerant)
-      $raise(Opal.NameError, 'uninitialized class variable '+name+' in '+module.$name());
-
-    return nil;
-  }
 
   function isRoot(proto) {
     return proto.hasOwnProperty('$$iclass') && proto.hasOwnProperty('$$root');
